Выносит форму в отдельный компонент
